chore: added comments and moved to dedicated folder
