+++ Source/Immutable/Private/Immutable/Transak/TransakWebBrowser.cpp
@@ -3,7 +3,7 @@
 #include "PlatformHttp.h"
 
 #include "Immutable/ImmutableUtilities.h"
-#include "Immutable/TransakConfig.h"
+#include "Immutable/Transak/TransakConfig.h"
 
 #if (ENGINE_MAJOR_VERSION >= 5 && ENGINE_MINOR_VERSION >= 1)
 #include "SWebBrowser.h"
